+++ test/helper.rb
@@ -10,14 +10,8 @@ require "minitest/mock_expectations"
 require "phenix"
 
 ENV["RAILS_ENV"] = "test"
-ENV["LEGACY_CONNECTION_HANDLING"] = "true" if ENV["LEGACY_CONNECTION_HANDLING"].nil?
 
-if ActiveRecord.version < Gem::Version.new("7.1")
-  ActiveRecord::Base.legacy_connection_handling = (ENV["LEGACY_CONNECTION_HANDLING"] == "true")
-end
-
-LEGACY_CONNECTION_HANDLING =
-  ActiveRecord.version < Gem::Version.new("7.1") && ActiveRecord::Base.legacy_connection_handling
+ActiveRecord::Base.legacy_connection_handling = false if ActiveRecord::Base.respond_to?(:legacy_connection_handling)
 
 ActiveRecord::Base.logger = Logger.new(File.dirname(__FILE__) + "/test.log")
 
@@ -50,53 +44,7 @@ module ARHPTestSetup
 
   def arhp_create_models
     return if ARHPTestSetup.const_defined?(:Pool1DbA)
-
-    if LEGACY_CONNECTION_HANDLING
-      eval(<<-RUBY, binding, __FILE__, __LINE__ + 1)
-        # The placement of the Pool1DbC class is important so that its
-        # connection will not be the most recent connection established
-        # for test_pool_1.
-        class Pool1DbC < ActiveRecord::Base
-          establish_connection(:test_pool_1_db_c)
-        end
-
-        class Pool1DbA < ActiveRecord::Base
-          self.table_name = "tests"
-          establish_connection(:test_pool_1_db_a)
-        end
-
-        class Pool1DbAOther < ActiveRecord::Base
-          self.table_name = "tests"
-          establish_connection(:test_pool_1_db_a)
-        end
-
-        class Pool1DbAReplica < ActiveRecord::Base
-          self.table_name = "tests"
-          establish_connection(:test_pool_1_db_a_replica)
-        end
-
-        class Pool1DbB < ActiveRecord::Base
-          self.table_name =  "tests"
-          establish_connection(:test_pool_1_db_b)
-        end
-
-        class Pool2DbD < ActiveRecord::Base
-          self.table_name = "tests"
-          establish_connection(:test_pool_2_db_d)
-        end
-
-        class Pool2DbE < ActiveRecord::Base
-          self.table_name = "tests"
-          establish_connection(:test_pool_2_db_e)
-        end
-
-        class Pool3DbE < ActiveRecord::Base
-          self.table_name = "tests"
-          establish_connection(:test_pool_3_db_e)
-        end
-      RUBY
-    else
-      eval(<<-RUBY, binding, __FILE__, __LINE__ + 1)
+    eval(<<-RUBY, binding, __FILE__, __LINE__ + 1)
         class AbstractPool1DbC < ActiveRecord::Base
           self.abstract_class = true
           connects_to database: { writing: :test_pool_1_db_c }
@@ -171,8 +119,7 @@ module ARHPTestSetup
         class ShardedModel < AbstractShardedModel
           self.table_name = "tests"
         end
-      RUBY
-    end
+    RUBY
   end
 
   def current_database(klass)
@@ -188,18 +135,5 @@ module ARHPTestSetup
     yield if block_given?
   ensure
     mod.define_method(method_name, method_body)
-  end
-
-  def simulate_rails_app_active_record_railties
-    if LEGACY_CONNECTION_HANDLING
-      # Necessary for testing ActiveRecord 6.0 which uses the connection
-      # handlers when clearing query caches across all handlers when
-      # an operation that dirties the cache is involved (e.g. create/insert,
-      # update, delete/destroy, truncate, etc.)
-      # In Rails 6.1 this is only present when legacy_connection_handling=true
-      ActiveRecord::Base.connection_handlers = {
-        ActiveRecord::Base.writing_role => ActiveRecord::Base.default_connection_handler
-      }
-    end
   end
 end