+++ test/test_arhp_wrong_db.rb
@@ -5,11 +5,7 @@ require_relative "helper"
 class ActiveRecordHostPoolWrongDBTest < Minitest::Test
   include ARHPTestSetup
   def setup
-    if LEGACY_CONNECTION_HANDLING
-      Phenix.load_database_config
-    else
-      Phenix.load_database_config "test/three_tier_database.yml"
-    end
+    Phenix.load_database_config "test/three_tier_database.yml"
   end
 
   def teardown